requirements.txt
allure-pytest==2.16.0
allure-python-commons==2.16.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
greenlet==3.5.0
idna==3.13
iniconfig==2.3.0
packaging==26.2
playwright==1.58.0
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
pytest==9.0.3
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dotenv==1.2.2
python-slugify==8.0.4
requests==2.33.1
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
